Port memfs for ARM.
