+++ kernel/fs_node/Cargo.toml
@@ -7,7 +7,6 @@ description = "defines the traits for File and Directory. These files and direct
 
 [dependencies]
 spin = "0.4.5"
-x86_64 = { path = "../../libs/x86_64" } # currently using our local copy, forked from Phil Opp's crate
 
 [dependencies.lazy_static]
 features = ["spin_no_std"]
