+++ kernel/memfs/Cargo.toml
@@ -7,19 +7,12 @@ description = "contains the MappedPages-backed implementation of the fs_node tra
 
 [dependencies]
 spin = "0.4.5"
-x86_64 = { path = "../../libs/x86_64" } # currently using our local copy, forked from Phil Opp's crate
 
 [dependencies.fs_node]
 path = "../fs_node"
 
 [dependencies.memory]
 path = "../memory"
-
-[dependencies.log]
-version = "0.4.8"
-
-[dependencies.irq_safety]
-git = "https://github.com/kevinaboos/irq_safety"
 
 [lib]
 crate-type = ["rlib"]
