+++ kernel/nano_core/Cargo.toml
@@ -8,23 +8,27 @@ build = "../../build.rs"
 [dependencies]
 cfg-if = "1.0.0"
 
+[dependencies.logger]
+path = "../logger"
+
+[dependencies.memory_initialization]
+path = "../memory_initialization"
+
+[dependencies.log]
+version = "0.4.8"
+
 [target.'cfg(target_arch = "x86_64")'.dependencies]
 spin = "0.4.10"
 multiboot2 = "0.7.1"
 # x86_64 = { git = "https://github.com/kevinaboos/x86_64" }
 x86_64 = { path = "../../libs/x86_64" } # currently using our local copy, forked from Phil Opp's crate
 
-[target.'cfg(target_arch = "x86_64")'.dependencies.log]
-version = "0.4.8"
 
 [target.'cfg(target_arch = "x86_64")'.dependencies.kernel_config]
 path = "../kernel_config"
 
 [target.'cfg(target_arch = "x86_64")'.dependencies.irq_safety]
 git = "https://github.com/kevinaboos/irq_safety"
-
-[target.'cfg(target_arch = "x86_64")'.dependencies.logger]
-path = "../logger"
 
 [target.'cfg(target_arch = "x86_64")'.dependencies.state_store]
 path = "../state_store"
@@ -50,18 +54,20 @@ path = "../panic_entry"
 [target.'cfg(target_arch = "x86_64")'.dependencies.vga_buffer]
 path = "../vga_buffer"
 
-[dependencies.memory_initialization]
-path = "../memory_initialization"
-
 [target.'cfg(target_arch = "arm")'.dependencies]
 cortex-m-rt = "0.6.13"
-cortex-m-semihosting = "0.3.7"
 panic-halt = "0.2.0"
 alloc-cortex-m = "0.4.1"
 
 [target.'cfg(target_arch = "arm")'.dependencies.cortex-m]
 features = ["inline-asm"]
 version = "0.6.4"
+
+[target.'cfg(target_arch = "arm")'.dependencies.memfs]
+path = "../memfs"
+
+[target.'cfg(target_arch = "arm")'.dependencies.root]
+path = "../root"
 
 
 [lib]
