+++ kernel/page_allocator/Cargo.toml
@@ -6,17 +6,23 @@ version = "0.1.0"
 build = "../../build.rs"
 
 [dependencies]
+cfg-if = "1.0.0"
 spin = "0.4.10"
 intrusive-collections = "0.9.0"
-
-[dependencies.log]
-version = "0.4.8"
 
 [dependencies.kernel_config]
 path = "../kernel_config"
 
 [dependencies.memory_structs]
 path = "../memory_structs"
+
+[target.'cfg(target_arch = "x86_64")'.dependencies.log]
+version = "0.4.8"
+
+[target.'cfg(target_arch = "arm")'.dependencies.lazy_static]
+features = ["spin_no_std", "nightly"]
+version = "1.2.0"
+
 
 [lib]
 crate-type = ["rlib"]
