+++ kernel/root/Cargo.toml
@@ -6,7 +6,6 @@ description = "a special concrete implementation of the Directory trait; differs
 
 [dependencies]
 spin = "0.4.5"
-x86_64 = { path = "../../libs/x86_64" } # currently using our local copy, forked from Phil Opp's crate
 
 [dependencies.lazy_static]
 features = ["spin_no_std"]
